\name{mk.pt.constr}
\alias{mk.pt.constr}
\title{COBS auxiliary for constructing pointwise constraint specifications}
\description{
  COBS (\code{\link{cobs}}) auxiliary function for constructing the
  pointwise constraint specification list from the \code{pointwise}
  3-column matrix (as used as argument in \code{\link{cobs}}).
}
\usage{
mk.pt.constr(pointwise)
}
\arguments{
  \item{pointwise}{numeric 3-column matrix, see \code{\link{cobs}}.}
}
\value{
  A list with components
%% FIXME
  \item{comp1 }{Description of 'comp1'}

}
\author{Martin Maechler}
\examples{
}
\keyword{misc}

